Update UploadAvatar and BasicInformation
